:recycle: refactor: rename the symbol function
BREAKING CHANGE: rename `nodeToSketchSymbol` to nodeToSymbol
